// backend/src/common/models/Purchase.ts
export interface Purchase {
  document_no: string;
  status: number;
  date: string;
  gross_total: number;
  notes?: string;
  document_type: string;
  pending_total: number;
  retention: any;
  settlement_total: number;
  due_date: string;
  emailed: any;
  tax_payable: number;
  net_total: number;
  document_hash_sum: any;
  hash_control: any;
  supplier_tax_registration_number: string;
  supplier_business_name: string;
  supplier_address_detail?: string;
  supplier_postcode: string;
  supplier_city?: string;
  created_at: string;
  updated_at: string;
  vehicle_registration: any;
  supplier_tax_country_region: string;
  supplier_country: string;
  printed: any;
  currency_conversion_rate: number;
  currency_iso_code: string;
  system_entry_date: string;
  retention_value: number;
  acts_as_shipment_document: any;
  shipment_loading_time: any;
  shipment_unloading_time: any;
  apply_retention_when_paid: any;
  pending_retention_value: number;
  communication_status?: string;
  shipment_country: any;
  base_currency_gross_total: any;
  base_currency_net_total: any;
  base_currency_pending_total: any;
  base_currency_settlement_total: any;
  base_currency_retention_total: any;
  communication_code: any;
  shipment_loading_address_detail?: string;
  shipment_loading_postcode?: string;
  shipment_loading_city?: string;
  shipment_loading_country: string;
  shipment_unloading_address_detail?: string;
  shipment_unloading_city: string;
  shipment_unloading_country: string;
  shipment_unloading_postcode: string;
  is_third_party: any;
  third_party_type: any;
  document_area: string;
  parent_document_area: any;
  communication_code_source: any;
  cashed_vat: any;
  retention_type: any;
  external_reference?: string;
  document_series_prefix: string;
  vat_included_prices: boolean;
  voided_reason?: string;
  settlement_with_tax: number;
  settlement_without_tax: number;
  line_settlement_with_tax: number;
  line_settlement_without_tax: number;
  total_pending_quantity: any;
  child_documents_ids: any;
  parent_documents_ids: any;
  payments_ids?: number[];
  operation_country: string;
  settlement_percentage: number;
  settlement_expression?: string;
  vat_percentage_red: any;
  vat_percentage_int: any;
  vat_percentage_nor: number;
  vat_incidence_ise_unrounded: number;
  vat_incidence_red_unrounded: number;
  vat_incidence_int_unrounded: number;
  vat_incidence_nor_unrounded: number;
  vat_total_red_unrounded: number;
  vat_total_int_unrounded: number;
  vat_total_nor_unrounded: number;
  settlement_with_tax_unrounded: number;
  settlement_without_tax_unrounded: number;
  line_settlement_with_tax_unrounded: number;
  line_settlement_without_tax_unrounded: number;
  retention_value_unrounded: number;
  base_currency_vat_incidence_nor: any;
  base_currency_vat_total_nor: any;
  base_currency_vat_incidence_int: any;
  base_currency_vat_total_int: any;
  base_currency_vat_incidence_red: any;
  base_currency_vat_total_red: any;
  base_currency_vat_incidence_ise: any;
  vat_incidence_ise: number;
  vat_incidence_red: number;
  vat_incidence_int: number;
  vat_incidence_nor: number;
  vat_total_red: number;
  vat_total_int: number;
  vat_total_nor: number;
  public_link: string;
  id: number;
  document_series_id: number;
  company_id: number;
  currency_id: number;
  supplier_id: number;
  user_id: number;
  lines: Line[];
}

export interface Line {
  item_type: string;
  item_id: number;
  item_code: string;
  description: string;
  quantity: number;
  unit_price: number;
  net_unit_price: number;
  item_unit_price_includes_vat?: boolean;
  settlement_amount: any;
  net_amount_with_header_settlement: number;
  amount: number;
  net_amount: number;
  tax_amount: number;
  retention_value: any;
  base_currency_unit_price: any;
  base_currency_net_unit_price: any;
  amount_with_header_settlement: number;
  base_currency_amount: any;
  base_currency_net_amount: any;
  base_currency_tax_amount: any;
  created_at: string;
  updated_at: string;
  original_document_no: any;
  settlement_with_tax: number;
  settlement_without_tax: number;
  date: any;
  pending_total?: number;
  pending_retention_value: any;
  pending_quantity?: number;
  imported_quantity: any;
  settlement_percentage: number;
  base_currency_settlement_with_tax: any;
  base_currency_settlement_without_tax: any;
  tax_code: string;
  tax_percentage: number;
  tax_country_region: string;
  settlement_expression: any;
  id: number;
  document_id: number;
  product_id: number;
  tax_id: number;
  unit_of_measure_id: number;
}